requirements/development.txt
ipython
requests
django-extensions
Werkzeug
pylint
pylint-django
django-stubs
djangorestframework-stubs
mysqlclient
